Made add tool into its own component and added helper functions
