+++ client/src/views/Home/Home.js
@@ -8,13 +8,8 @@ import TextField from '@material-ui/core/TextField';
 import InputAdornment from '@material-ui/core/InputAdornment';
 import SearchRounded from '@material-ui/icons/SearchRounded';
 import ToolDetailPopup from '../../components/ToolDetailPopup';
+import AddToolPopup from '../../components/AddToolPopup';
 import Button from '@material-ui/core/Button';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import { makeStyles } from '@material-ui/core/styles';
 
 // Styled components
 const RootContainer = styled.div`
@@ -36,7 +31,7 @@ const ToolImage = styled.img`
   }
 `;
 
-const dummyData = [
+let dummyData = [
   {
     src: logo,
     keywords: [
@@ -64,39 +59,44 @@ const dummyData = [
   }
 ];
 
-const useStyles = makeStyles(theme => ({
-  root: {
-    '& > *': {
-      margin: theme.spacing(25),
-    },
-  },
-  input: {
-    display: 'none'
-  },
-}));
-
 function Home() {
   const [filteredTools, setFilteredTools] = useState(dummyData);
   const [selectedTool, setSelectedTool] = useState(null);
-  const [open,setOpen] = React.useState(false);
 
+  // Used to control opening up popup of adding a tool
+  const [open,setOpen] = React.useState(false);
   const handleClickOpen = () => {
     setOpen(true);
   }
-
   const handleClose = () => {
     setOpen(false);
   };
-  
   const addToDummy = () => {
     let obj = {};
     obj['src'] = logo;
     obj['keywords'] = ['testing if this works', 'abc'];
-    dummyData.push(obj);
+    //for (let i = 0; i < 2000; i++) {
+      dummyData.push(obj);
+    //}
     setOpen(false);
   };
+  // End of control add tool (will come back to clean this area too, AddToolPopup component is now working)
 
-  const classes = useStyles();
+  // Delete a tool (in the future use mongoose ID)
+  const deleteFromDummy = (data) => {
+    //console.log("inside of delete from dummy!!!");
+    //console.log("the data is: ", {data});
+    const filterItems = dummyData.filter(function(item) {
+      //console.log(item);
+      return item !== data;
+    });
+    //console.log(filterItems);
+    dummyData = filterItems;
+    setFilteredTools(dummyData);
+    // Small hack to clear search field, in textfield "id='search'"
+    document.getElementById('search').value = ''
+  };
+  // End of delete
 
   const filterTools = e => {
     const searchTerm = e.target.value;
@@ -125,6 +125,7 @@ function Home() {
           autoCapitalize='off'
           autoComplete='off'
           spellCheck='false'
+          id='search'
           fullWidth
           InputProps={{
             startAdornment: (
@@ -139,44 +140,11 @@ function Home() {
       <Button variant="contained" color="primary" onClick={handleClickOpen}>
         Upload A New Tool
       </Button>
-      <Dialog open={open} onClose={handleClose} aria-labelledby="form-dialog-title"> 
-        <DialogTitle id="form-dialog-title" >Upload A New Tool</DialogTitle >
-        <DialogContent>
-          <DialogContentText>
-            To upload a new tool, fill in the info below. (Just click upload for demo)
-          </DialogContentText>
-          <div className={classes.root}>
-            <input
-              accept="image/*"
-              className={classes.input}
-              id="contained-button-file"
-              type="file"
-            />
-            <label htmlFor="contained-button-file">
-              <Button variant="contained" color="primary" component="span" >
-                Upload An Image
-              </Button>
-            </label>
-          </div>
-          <TextField
-            autoFocus
-            margin="dense"
-            id="keywords"
-            label="Keywords"
-            type="keywords"
-            fullWidth
-          />
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose} color="primary">
-            Cancel
-          </Button>
-          <Button onClick={addToDummy} color="primary">
-            Upload
-          </Button>
-        </DialogActions>
-      </Dialog>
-
+      <AddToolPopup
+        open={open}
+        handleClose={handleClose}
+        addToDummy={addToDummy}
+      />
       <Grid container>
         {filteredTools.map((d, index) => (
           <Grid item xs={4} key={index}>
@@ -188,6 +156,7 @@ function Home() {
         tool={selectedTool}
         isOpen={selectedTool !== null}
         close={() => setSelectedTool(null)}
+        deleteFromDummy={deleteFromDummy}
       />
     </RootContainer>
   );
